+++ CHANGELOG.md
@@ -112,6 +112,7 @@
 - BEIS users only see an organisation's activities on that organisation's show page (bug)
 - XML file for projects now shows the identifiers for the parent activities.
 - Fix descriptive labels on action links
+- Sector step is made up of two questions sector category and sector
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-4...HEAD
 [release-4]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-3...release-4